+++ src/runtime/tensormap_and_ringbuffer/host/runtime_maker.cpp
@@ -216,7 +216,7 @@ extern "C" int init_runtime_impl(Runtime *runtime,
     runtime->set_pto2_gm_heap(gm_heap);
 
     // Allocate PTO2 shared memory
-    int32_t sm_size = pto2_sm_calculate_size(PTO2_TASK_WINDOW_SIZE, PTO2_DEP_LIST_POOL_SIZE);
+    size_t sm_size = pto2_sm_calculate_size(PTO2_TASK_WINDOW_SIZE, PTO2_DEP_LIST_POOL_SIZE);
     void* sm_ptr = runtime->host_api.device_malloc(static_cast<size_t>(sm_size));
     if (sm_ptr == nullptr) {
         std::cerr << "Error: Failed to allocate PTO2 shared memory\n";
@@ -263,7 +263,7 @@ extern "C" int validate_runtime_impl(Runtime *runtime) {
     // PTO2 (device orchestration): graph output may be in packed buffer
     void* pto2_sm = runtime->get_pto2_gm_sm_ptr();
     uint64_t graph_out_ptr = 0;
-    int32_t graph_out_size = 0;
+    size_t graph_out_size = 0;
 
     if (pto2_sm != nullptr) {
         // Copy header from device to host to read graph_output_ptr/size
